Fix: support large file sizes on 32 bit archs (#467)
Fixes the following error message:

```
cannot unmarshal number into Go struct field
ExifInfo.Assets.items.exifInfo.fileSizeInByte of type int
```

This is caused by a video in immich that is larger than can fit into an
int32.

+++ cmd/upload/assets.go
@@ -46,7 +46,7 @@ func (ai *AssetIndex) AddLocalAsset(la *browser.LocalAssetFile, immichID string)
 		DeviceAssetID:    la.DeviceAssetID(),
 		OriginalFileName: strings.TrimSuffix(path.Base(la.Title), path.Ext(la.Title)),
 		ExifInfo: immich.ExifInfo{
-			FileSizeInByte:   int(la.Size()),
+			FileSizeInByte:   int64(la.Size()),
 			DateTimeOriginal: immich.ImmichTime{Time: la.Metadata.DateTaken},
 			Latitude:         la.Metadata.Latitude,
 			Longitude:        la.Metadata.Longitude,
